Step 2: rename infertsconfig file back with correct capitalization
